pyproject.toml
[project]
name = "qog-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "neo4j>=5.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
qog-dashboard = "main:run"
